refactor(config): remove dependencies on adapters/models
Remove config layer's dependencies on adapters and models to establish
clean layer boundaries.

Changes:
- Move BranchConvention from models/ to config/ (pure dataclass)
- Update models/branch_convention.py to re-export from config (backward compat)
- Replace top-level adapters import in profile_config.py with lazy import
- Use TYPE_CHECKING for AdapterManifest type annotation
- Update test imports to use config.branch_convention

Verification:
- All config tests pass (80 tests)
- BranchConvention re-export tests pass (5 tests)
- IssueFlowService tests pass (12 tests)
- No type errors (mypy)
- No lint errors (ruff)
- No top-level imports from adapters/models in config layer
